General/chatbot_test/requirements.txt
Flask == 3.0.0
requests == 2.31.0
python-dotenv
pymessenger == 0.0.7.0